feat(service): validaciones a la hora de eliminar un objeto ya sea especie u zona

+++ src/main/java/com/playko/zoologico/service/impl/ZonaService.java
@@ -2,11 +2,14 @@ package com.playko.zoologico.service.impl;
 
 import com.playko.zoologico.dto.request.ZonaRequestDto;
 import com.playko.zoologico.dto.response.ZonaResponseDto;
+import com.playko.zoologico.entity.Animal;
 import com.playko.zoologico.entity.Especie;
 import com.playko.zoologico.entity.Zona;
 import com.playko.zoologico.exception.NoDataFoundException;
+import com.playko.zoologico.exception.animal.ZonaConAnimalesException;
 import com.playko.zoologico.exception.zona.ZonaAlreadyExistsException;
 import com.playko.zoologico.exception.zona.ZonaNotFoundException;
+import com.playko.zoologico.repository.IAnimalRepository;
 import com.playko.zoologico.repository.IZonaRepository;
 import com.playko.zoologico.service.IZonaService;
 import jakarta.transaction.Transactional;
@@ -15,12 +18,14 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class ZonaService implements IZonaService {
     private final IZonaRepository zonaRepository;
+    private final IAnimalRepository animalRepository;
     @Override
     public ZonaResponseDto obtenerZonaPorId(Long id) {
         Zona zona = zonaRepository.findById(id)
@@ -70,18 +75,43 @@ public class ZonaService implements IZonaService {
     public void eliminarZona(Long id) {
         Zona zona = zonaRepository.findById(id)
                 .orElseThrow(ZonaNotFoundException::new);
+
+        boolean hayAnimales = animalRepository.existsByEspecie_Zona(zona);
+        if (hayAnimales) {
+            throw new ZonaConAnimalesException();
+        }
+
         zonaRepository.delete(zona);
     }
 
     private ZonaResponseDto mapToResponseDto(Zona zona) {
-        ZonaResponseDto dto = new ZonaResponseDto();
-        dto.setId(zona.getId());
-        dto.setNombre(zona.getNombre());
-        List<String> nombresEspecies = zona.getEspecies()
+        List<String> nombresEspecies = zona.getEspecies() != null
+                ? zona.getEspecies()
                 .stream()
                 .map(Especie::getNombre)
-                .toList();
-        dto.setEspecies(nombresEspecies);
-        return dto;
+                .toList()
+                : List.of();
+
+        List<String> nombresAnimales = zona.getEspecies() != null
+                ? zona.getEspecies()
+                .stream()
+                .flatMap(especie -> {
+                    if (especie.getAnimales() == null) {
+                        return Stream.<String>empty();
+                    }
+                    return especie.getAnimales()
+                            .stream()
+                            .map(Animal::getNombre);
+                })
+                .toList()
+                : List.of();
+
+
+        return new ZonaResponseDto(
+                zona.getId(),
+                zona.getNombre(),
+                nombresEspecies,
+                nombresAnimales
+        );
     }
 }
